#58 added files for GATTAnaerobicHeartRateLowerLimit and GATTAnaerobicHeartRateUpperLimit
